{{cookiecutter.repo_name}}/requirements/base.txt
Django==3.1.12
Pillow==7.2.0
{% if cookiecutter.postgres == "y" or cookiecutter.postgres == "Y" %}
psycopg2-binary==2.8.6
{% endif %}
djangorestframework==3.11.1
drf-yasg==1.17.1
django-filter==2.3.0
django-cors-headers==3.7.0
django-extensions==3.0.4
django-constance[database]==2.7.0
black==20.8b1
flake8~=3.9.0
